refactor(agents): preserve HTTP runtime seam
